fix: stop _adaptToScreenSizeChange from being called because of mobile scroll
Scrolling on mobile devices triggers the 'resize' event listener. This happens because the address bar pops in and out and that effects the viewport height. In turn this calls the _adaptToScreenSizeChange method, resetting the testimonials component every time a user scrolls. To fix this we check if the width of the viewport was affected before calling _adaptToScreenSizeChange.